feat(cli): clean main function with separate handles and format project
