+++ aether-cli/src/config.rs
@@ -1,4 +1,7 @@
-use std::{collections::HashMap, fs::OpenOptions, io::Write, path::PathBuf};
+use std::collections::HashMap;
+use std::fs::OpenOptions;
+use std::io::Write;
+use std::path::PathBuf;
 
 use serde::{Deserialize, Serialize};
